feat: Animal Script
각 동물들의 부모 클래스가 되는 Animal Script 추가
