+++ Assets/RHU/Scripts/Animal/Animal.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6b6c201e352759f4986f45de734dfe35
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
